commit de avl para busquedas

+++ utils/Parser.cpp
@@ -22,8 +22,8 @@ std::vector<DataTypeInfo> DataLoader::parseTypes(const std::string& typeLine) {
             types.emplace_back(CHAR);
         }
         else if (typeStr.find("varchar") != std::string::npos) {
-            size_t openParen = typeStr.find('(');
-            size_t closeParen = typeStr.find(')');
+            unsigned int openParen = typeStr.find('(');
+            unsigned int closeParen = typeStr.find(')');
             if (openParen == std::string::npos || closeParen == std::string::npos) {
                 throw std::invalid_argument("Formato varchar inválido: " + typeStr);
             }
@@ -57,7 +57,7 @@ std::vector<Registro> DataLoader::loadFromFile(const std::string& filename, Data
             std::istringstream iss(line);
             Registro registro;
             cout<<"Numero de atributos: "<<types.size()<<endl;            
-            for (size_t i = 0; i < types.size(); ++i) {
+            for (unsigned int i = 0; i < types.size(); ++i) {
                 std::string value;
                 if (!(iss >> value)) {
                     throw std::runtime_error("Formato de registro inválido");
